forgot delete event on private channel for charasheet

+++ src/events/on_delete.ts
@@ -14,18 +14,12 @@ export const delete_channel = (client	: EClient): void => {
 			//search channelID in database and delete it
 			const guildID = channel.guild.id;
 			const db = client.settings;
-			if (db.get(guildID, "templateID.channelId") === channelID) {
-				db.delete(guildID, "templateID");
-			}
-			if (db.get(guildID, "logs") === channelID) {
-				db.delete(guildID, "logs");
-			}
-			if (db.get(guildID, "managerId") === channelID) {
-				db.delete(guildID, "managerId");
-			}
-			if (db.get(guildID, "rollChannel") === channelID) {
-				db.delete(guildID, "rollChannel");
-			}
+			if (db.get(guildID, "templateID.channelId") === channelID) db.delete(guildID, "templateID");
+			if (db.get(guildID, "logs") === channelID) db.delete(guildID, "logs");
+			if (db.get(guildID, "managerId") === channelID) db.delete(guildID, "managerId");
+			if (db.get(guildID, "privateChannel") === channelID) db.delete(guildID, "privateChannel");
+			if (db.get(guildID, "rollChannel") === channelID) db.delete(guildID, "rollChannel");
+			
 		} catch (error) {
 			err(error);
 			if (channel.isDMBased()) return;
@@ -41,12 +35,12 @@ export const delete_thread = (client: EClient): void => {
 			//search channelID in database and delete it
 			const guildID = thread.guild.id;
 			const db = client.settings;
-			if ((thread.name === "📝 • [STATS]" && thread.parentId === db.get(guildID, "templateID.channelId")) || thread.id === db.get(guildID, "templateID.managerId")) {
+			if ((thread.name === "📝 • [STATS]" && thread.parentId === db.get(guildID, "template.channelId")) || thread.id === db.get(guildID, "managerId") || thread.id === db.get(guildID, "privateChannel")) {
 				//verify if the user message was in the thread
 				cleanUserDB(db, thread);
 			}
 			if (db.get(guildID, "logs") === thread.id) db.delete(guildID, "logs");
-			if (db.get(guildID, "templateID.channelId") === thread.id) db.delete(guildID, "templateID");
+			if (db.get(guildID, "template.channelId") === thread.id) db.delete(guildID, "template");
 		} catch (error) {
 			err(error);
 			if (thread.isDMBased()) return;
@@ -62,10 +56,15 @@ export const delete_message = (client: EClient): void => {
 			const messageId = message.id;
 			//search channelID in database and delete it
 			const guildID = message.guild.id;
-			
-			if (client.settings.get(guildID, "templateID.messageId") === messageId) {
-				client.settings.delete(guildID, "templateID");
-			}
+			console.log(message.channel.id, message.thread?.id);
+			const channel = message.channel;
+			if (channel.isDMBased()) return;
+			if (client.settings.get(guildID, "templateID.messageId") === messageId) client.settings.delete(guildID, "templateID");
+			if (
+				channel.id !== client.settings.get(guildID, "privateChannel") 
+				|| channel.id !== client.settings.get(guildID, "managerId") ||
+				channel.name !== "📝 • [STATS]"
+			) return;
 			const dbUser = client.settings.get(guildID, "user");
 			if (dbUser && Object.keys(dbUser).length > 0){
 				for (const [user, values] of Object.entries(dbUser)) {
